requirements.txt
streamlit>=1.36
transformers>=4.41
tokenizers>=0.15
sentence-transformers>=2.7 ; platform_system != "Windows" or python_version < "3.13"
faiss-cpu>=1.7
scikit-learn>=1.4
numpy>=1.26
scipy>=1.11
pdfplumber>=0.11
pypdf>=4.2
python-docx>=1.1